Extract SessionDisplay mixins to session package
Split the large session_window implementation into a session package and utility modules: added session/constants.py (BREAK_IMAGE_PATH, supported image sets, sound_file), and new mixin modules session/image_mods.py, session/shortcuts.py, session/timer.py, and session/zoom_pan.py. SessionDisplay now composes those mixins (cleaning up a lot of inline logic) and the session package has a small __init__ docstring. Also moved config and update-checker helpers into gesturesesh.utils (utils/config.py, utils/update_checker.py) and updated imports across app modules and tests accordingly. This refactor reduces circular imports and separates concerns for easier maintenance and testing.
