+++ src/gesturesesh/session/__init__.py
@@ -1,0 +1,1 @@
+"""Behavior mixins for the SessionDisplay slideshow window."""
